Code cleanup

+++ src/storagekey.cpp
@@ -35,7 +35,7 @@ std::ostream &operator<<(std::ostream &os, const StorageKey &key) {
     std::cout << "MK dump:\t";
     std::ios state(nullptr);
     state.copyfmt(std::cout);
-    for (int i = 0; i < key.getKeySize(); i++) {
+    for (size_t i = 0; i < key.getKeySize(); i++) {
         if (i && !(i % 16))
             std::cout << ("\n\t\t");
         std::cout << std::hex << std::setfill('0') << std::setw(2) << (int) (unsigned char) key.getKey()[i] << " ";
